Update Spine version to `1.1.2`
Also cleaned up the remaining SLF4J dependencies.

The `firebase-web` needs to keep an SLF4J binding though, as it's
required by the Firebase Admin SDK. The SLF4J version is added to
the local `version.gradle` and does not rely on `config`, so any
trace of SFL4J can be removed `config` when the `1.1.2` release is
done.