Split entity and event subscriptions
Event subscription now returns an `EventSubscriptionObject`.
The `EventSubscriptionObject` does not have redundant `itemChanged`
and `itemRemoved` callbacks and instead holds a single observable
which is named `eventEmitted` and is of `Observable<Event>` type.

The `SubscriptionBridge` now writes `Event` instances to the Firebase
database instead of event messages to allow this behavior.