Store empty subscription updates as real `null`s in Firebase database
Previously, we added the node with the string value of `"null"` to the database on empty subscription update. This led to the issue when the `itemChanged` instead of `itemRemoved` callback was triggered on the client side when the entity state became no longer matching the subscription criteria.

Now, the real `null` is stored to the database on empty subscription update, which in terms of Firebase means the child node is deleted. Thus, the "correct" `itemRemoved` callback is triggered.

Also, a small bug with trying to store the `null` values with the `ImmutableMap` collector is fixed.