Fix issue with `null` grand origin specified to the filter
The grand origin in the command context is never actually `null` on the server. It either has some value or equals to the default instance.

So for the filter value to equal the value in the corresponding command context on the server, we need to set the origin in a same way.