Fix the tests
Due to what is seemingly an issue with the Spine core the `MirrorProjection` messages can be delivered only within a single `BoundedContext` - the one which is registered last. So the "main" bounded context, where the multitenant tests run and multiple aggregates are read, is now registered last and the "auxiliary" single-tenant context will contain a `UserInfo` aggregate and a corresponding projection instead of having a single visible aggregate.

Also fixed how the `observe`d event is handled in the single-tenant tests.